chore: small improvements
